edit yield type in gen type hint

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -6,7 +6,7 @@ import asyncio
 from typing import Generator
 
 
-async def async_generator() -> Generator[int, None, None]:
+async def async_generator() -> Generator[float, None, None]:
     """loop 10 times, wait asynchronously then yield a random number"""
     for _ in range(10):
         await asyncio.sleep(1)
